Updated User Creation form + Lang change
Updated the User creation form to display you have X licenses left to
allocate instead of X out of X. I also changed the string for
user_courses_for which makes more sense in a commit I am going to post
soon.

This works 100% but I will be tidying it up in the following week.