New Company User Licenses Edit Page
This is a complete overhaul of the company_users_licenses_from. It  has
been redesigned, new strings included, a help blurb, Reformatted the
information displayed and also implemented a check to prevent people
trying to allocate/unallocate without selecting a license. Also included
a better error message for users with no active licenses.