lib.php user informed if certificate has been generated
Previously, the function xmldb_local_iomad_track_record_certificates would run, and nothing else would happen. However, if it does not generate a certificate then the user is none the wiser. This fix is an addition to some code that has been modified in the declaration of the xmldb_local_iomad_track_record_certificates function whereby it will return true or false depending on the outcome of the certificate generation. For the certificate to generate, the user will need to have a timecompleted value in the local_iomad_track table. If they do not, the function will return false and the user will be informed.